wip: continue implementation of GameService and GameState
